refactor(plugin-worker-mailer): extract shared smtp helpers
